Add Habit component
