+++ src/components/List/List.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import styled from 'styled-components/macro'
-import Habit from '../Habit/HabitItem'
+import Habit from '../Habit/Habit'
 
 export default function List({ habits, onItemClick, displayToday }) {
   return (
@@ -8,16 +8,13 @@ export default function List({ habits, onItemClick, displayToday }) {
       <CurrentDate>{displayToday()}</CurrentDate>
       <HabitList>
         {habits.map((habit) => (
-          <div key={habit.id}>
-            <HabitContainer>
-              <Habit
-                id={habit.id}
-                name={habit.name}
-                completed={habit.completed}
-                onItemClick={onItemClick}
-              />
-            </HabitContainer>
-          </div>
+          <Habit
+            key={habit.id}
+            id={habit.id}
+            name={habit.name}
+            completed={habit.completed}
+            onItemClick={onItemClick}
+          />
         ))}
       </HabitList>
     </>
@@ -34,12 +31,4 @@ const HabitList = styled.section`
   margin-top: 20px;
   position: relative;
   text-align: center;
-`
-const HabitContainer = styled.div`
-  display: inline-block;
-  margin-top: 10px;
-  width: 260px;
-  height: 25px;
-  background-color: #ffc179;
-  border-radius: 10px;
 `
